+++ webpack.config.js
@@ -33,7 +33,7 @@ module.exports = {
 
     path: path.resolve(__dirname, 'dist'),
     clean: true,
-    publicPath: '/',
+    publicPath: './',
 
   },
